backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
sqlalchemy==2.0.41
pydantic==2.11.7
pydantic-settings==2.9.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pillow==11.2.1 